requirements.txt
beautifulsoup4
requests
pandas
pymysql
python-dotenv
scrapy
schedule
